[INTERNAL][E] Move "Create Account" button in page
