+++ eclipse/src/saros/ui/messages.properties
@@ -111,6 +111,7 @@ PersonalizationPreferencePage_visible_information_group_text=Visual Information
 
 ConfigurationWizard_account_created_text=Your XMPP Account has already been created\n and added to your Saros configuration.\n\n Do you really want to cancel?
 ConfigurationWizard_account_created=Account Created
+ConfigurationWizard_button_create_account=Create Account...
 ConfigurationWizard_info_already_created_account=If you already have an XMPP account you can use it right away.\nOtherwise you can register an account at our website:\n<a href="https://saros-con.imp.fu-berlin.de:5280/register/new">https://saros-con.imp.fu-berlin.de:5280/register/new</a>
 
 ConfigurationSettingsWizardPage_0=Searching for gateways...
